refactor: modernize to 3.10+ idioms, enforce with ruff, split the big files
Tooling first: pyproject now carries a ruff config (E/W/F/I/UP/B/SIM/C4/
RET/PTH/PIE/PERF/RUF, line-length 120) and a PEP 735 dev group, so "modern
rules" is a thing the repo states and checks rather than a style anyone has
to remember. `ruff check remanga/` went 1629 -> 0.

Typing modernized to what requires-python already promises: PEP 585
(List/Dict/Tuple -> list/dict/tuple, 355 sites), PEP 604 (Optional[X] ->
X | None, 117), typing.Callable -> collections.abc, and one import order
across all 178 modules. 159 of them already had `from __future__ import
annotations`; the 9 that didn't are re-export __init__ files with no
annotations, so nothing here is only valid as a string.

Real fixes rather than cosmetics, where ruff found them:
  - zip(strict=True) at 4 sites whose operands are same-length by
    construction, so a broken invariant fails loudly instead of silently
    truncating to the shorter one
  - a dead `gen = PanelSheetGenerator` in _render_info_sheet (the identical
    line in create_panel_sheets is live and stayed)
  - list(map.values())[0] -> next(iter(...)), int(round(x)) -> round(x),
    stdout=PIPE/stderr=PIPE -> capture_output=True, append-loops ->
    comprehensions/extend, try/except/pass -> contextlib.suppress (every
    explanatory comment kept), os.path -> pathlib, ambiguous `l` -> `left`

Files split where there was a real seam, not to hit a number:
  - commands/registry.py 536 -> 39, into catalog/{setup,chapter,project}.py
    plus categories.py, setup_rows.py and help_text.py. Adding a command is
    still one edit in one place - now the file for its category, which is
    not a judgement call because the command already declares one.
  - status/panel.py: the same done/missing/pending/off ternary was inlined
    a dozen times with its markup written out fresh each time; that
    vocabulary is now status/badges.py
  - audio/tts.py 394 -> 316, clip gain/fades/atomic-export -> audio/clips.py
  - config/tts.py 333 -> 283, the engine catalogue -> config/tts_engines.py

No file is over 400 lines now (was one at 536); 143 of 178 are under 150,
mean 100. The remaining 300-400 line files are cohesive and were left
alone. Every moved symbol is still importable from its original path.

Verified, since there is no test suite:
  - all 498 resolved type signatures identical before/after via
    typing.get_type_hints (Optional[List[str]] and list[str] | None resolve
    to the same object, so this checks meaning, not spelling)
  - every line-wrap and string-split accepted only if the file's AST was
    unchanged, so no reformatting could alter a rendered string
  - `remanga --help` and all 23 subcommand --help outputs byte-identical
  - the full command registry - names, params, help, setup rows, ordering -
    identical apart from two lambdas' __qualname__ (they moved module)
  - status panel byte-identical across 5 chapters
  - config.json round-trip and the legacy flat-TTS migration unchanged
  - extracted clip helpers differential-tested against the originals over
    25 real clips x 4 fade settings x 3 gain settings: identical
  - 168 modules import clean; compileall clean

+++ remanga/humanize.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 
 def fmt_duration(seconds: float) -> str:
     """Seconds as "1h02m03s" / "2m03s" - hours only when there are any."""
-    seconds = int(round(seconds))
+    seconds = round(seconds)
     hours, rem = divmod(seconds, 3600)
     minutes, secs = divmod(rem, 60)
     return f"{hours}h{minutes:02d}m{secs:02d}s" if hours else f"{minutes}m{secs:02d}s"
